chore(path): move sportsIcons.tsx files to components directory
